+++ header/Firm.h
@@ -19,43 +19,6 @@ class Firm;
 class Producer;
 class Society;
 
-struct Plan {
-	// independent/input fields
-	Order * order;
-    Firm * firm;
-	std::vector<Person *> workers;
-    unsigned int local_work_hours_daily;
-
-	// dependent/output fields	
-	double predicted_turnaround_time;
-    double machinery_cost;
-    double labor_hours;
-    double raw_materials;
-    double total_hours;
-    double prd;
-    double labor_hours_remaining;
-    double raw_materials_remaining;
-    double total_hours_remaining;
-    double quantity_remaining;
-	int outgoing_units_consumed;
-};
-
-struct Order {
-    enum OrderStatus { ORDER_REQUESTED, ORDER_REJECTED, ORDER_IN_PROGRESS, ORDER_FINISHED };
-    Product * product;
-    int quantity;
-    Firm * customer;
-    double requested_turnaround_time;
-    OrderStatus status;
-    
-    Order(
-            Product * product,
-            int quantity,
-            Firm * customer,
-            double requested_turnaround_time
-    );
-};
-
 class Firm : public Agent {
   public:
     Firm();
