+++ src/Producer.cpp
@@ -7,7 +7,9 @@
 #include "Distributor.h"
 #include "Good.h"
 #include "Logger.h"
+#include "Order.h"
 #include "Person.h"
+#include "Plan.h"
 #include "Producer.h"
 #include "Product.h"
 #include "Sim.h"
